+++ database/migrations/2022_05_18_070306_create_designations_table.php
@@ -13,16 +13,19 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('designations', function (Blueprint $table) {
-            $table->id();
-            $table->integer('department_id')->nullable();
-            $table->integer('employer_id')->nullable();
-            $table->integer('activity_by')->nullable();    
-            $table->string('name')->nullable();
-            $table->integer('is_active')->default('1');
-            $table->integer('is_deleted')->default('0');
-            $table->timestamps();
-        });
+        // Check if the designations table exists
+        if (!Schema::hasTable('designations')) {
+            Schema::create('designations', function (Blueprint $table) {
+                $table->id();
+                $table->integer('department_id')->nullable();
+                $table->integer('employer_id')->nullable();
+                $table->integer('activity_by')->nullable();
+                $table->string('name')->nullable();
+                $table->integer('is_active')->default(1); // Changed default to integer value
+                $table->integer('is_deleted')->default(0); // Changed default to integer value
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -32,6 +35,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('designations');
+        // Check if the designations table exists before attempting to drop it
+        if (Schema::hasTable('designations')) {
+            Schema::dropIfExists('designations');
+        }
     }
 };
