+++ database/migrations/2022_05_20_085945_create_user_permission_actions_table.php
@@ -13,16 +13,19 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('user_permission_actions', function (Blueprint $table) {
-            $table->id();
-            $table->integer('user_id')->nullable();
-            $table->integer('user_permission_id')->nullable();
-            $table->integer('admin_module_id')->nullable();    
-            $table->integer('admin_sub_module_id')->nullable();
-            $table->integer('admin_module_action_id')->nullable();
-            $table->integer('is_active')->nullable(); 
-            $table->timestamps();
-        });
+        // Check if the user_permission_actions table exists
+        if (!Schema::hasTable('user_permission_actions')) {
+            Schema::create('user_permission_actions', function (Blueprint $table) {
+                $table->id();
+                $table->integer('user_id')->nullable();
+                $table->integer('user_permission_id')->nullable();
+                $table->integer('admin_module_id')->nullable();
+                $table->integer('admin_sub_module_id')->nullable();
+                $table->integer('admin_module_action_id')->nullable();
+                $table->integer('is_active')->default(1); // Added default value
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -32,6 +35,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('user_permission_actions');
+        // Check if the user_permission_actions table exists before attempting to drop it
+        if (Schema::hasTable('user_permission_actions')) {
+            Schema::dropIfExists('user_permission_actions');
+        }
     }
 };
