+++ database/migrations/2022_05_31_110253_create_blogs_table.php
@@ -13,16 +13,19 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('blogs', function (Blueprint $table) {
-            $table->id();
-            $table->string('name');
-            $table->string('page_name');
-            $table->string('image');
-            $table->text('body')->nullable();
-            $table->integer('is_active')->default('1');
-            $table->integer('is_delete')->default('0');
-            $table->timestamps();
-        });
+        // Check if the table already exists
+        if (!Schema::hasTable('blogs')) {
+            Schema::create('blogs', function (Blueprint $table) {
+                $table->id();
+                $table->string('name');
+                $table->string('page_name');
+                $table->string('image');
+                $table->text('body')->nullable();
+                $table->integer('is_active')->default('1');
+                $table->integer('is_delete')->default('0');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -32,6 +35,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('blogs');
+        // Optionally, drop the table if it exists
+        if (Schema::hasTable('blogs')) {
+            Schema::dropIfExists('blogs');
+        }
     }
 };
