+++ database/migrations/2023_04_05_095503_create_projects_table.php
@@ -13,18 +13,21 @@ return new class extends Migration
      */
     public function up()
     {
-        Schema::create('projects', function (Blueprint $table) {
-            $table->id();
-            $table->string('project_name',200);
-            $table->string('project_size',200);
-            $table->string('location',300);
-            $table->string('property_dimension',300);
-            $table->text('description');
-            $table->text('project_terms');
-            $table->enum('verification_status',['Pending','Approved','Rejected'])->default('Pending');
-            $table->enum('deleted',['No','Yes'])->default('No');
-            $table->timestamps();
-        });
+        // Check if the table already exists
+        if (!Schema::hasTable('projects')) {
+            Schema::create('projects', function (Blueprint $table) {
+                $table->id();
+                $table->string('project_name', 200);
+                $table->string('project_size', 200);
+                $table->string('location', 300);
+                $table->string('property_dimension', 300);
+                $table->text('description');
+                $table->text('project_terms');
+                $table->enum('verification_status', ['Pending', 'Approved', 'Rejected'])->default('Pending');
+                $table->enum('deleted', ['No', 'Yes'])->default('No');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
@@ -34,6 +37,9 @@ return new class extends Migration
      */
     public function down()
     {
-        Schema::dropIfExists('projects');
+        // Optionally, drop the table if it exists
+        if (Schema::hasTable('projects')) {
+            Schema::dropIfExists('projects');
+        }
     }
 };
